+++ packages/block-editor/src/components/colors/test/with-colors.js
@@ -29,7 +29,7 @@ describe( 'createCustomColorsHOC', () => {
 				} ),
 				setBackgroundColor: expect.any( Function ),
 			} ),
-			expect.anything()
+			undefined
 		);
 	} );