+++ packages/compose/src/hooks/use-disabled/index.ts
@@ -50,8 +50,8 @@ export default function useDisabled( {
 					if ( ! ( child instanceof defaultView.HTMLElement ) ) {
 						return;
 					}
-					if ( ! child.getAttribute( 'inert' ) ) {
-						child.setAttribute( 'inert', 'true' );
+					if ( ! child.hasAttribute( 'inert' ) ) {
+						child.setAttribute( 'inert', '' );
 						updates.push( () => {
 							child.removeAttribute( 'inert' );
 						} );