+++ packages/grid/src/dashboard-grid/grid-item.tsx
@@ -143,7 +143,7 @@ export function GridItem( {
 				>
 					<div
 						style={ { display: 'contents' } }
-						{ ...( dragging ? { inert: '' } : {} ) }
+						inert={ dragging || undefined }
 					>
 						{ actionableArea }
 					</div>